// src/app/workflow/models/contractApproval/contractApproval.interface.ts
export interface PhonesData{
    number: string;
    changeType: string;
    phoneType: string;
    country: string;
    areaCode: string;
    phone: string;
    phoneNotification?: string;
}

export interface EmailsData{
    number: string;
    changeType: string;
    email: string;
    emailNotification?: string;
}